+++ clientside/plugins/rotations/deathkight/frost/v4/01_amstaffix_deathknight_frost_loader.lua
@@ -48,6 +48,8 @@ local Config = {
     minEnemiesCountToRaiseDead = 2,
 
     minHPToUseDeathPact = 80,
+    --- IF HP < N, then use death pact instantly
+    useDeathPactMinHpForcibly = 40,
 
     useHowlingBlastMinEnemies = 3,